[workflows] added output directory option for type-stub generator

+++ src/main/scala/io/joern/typestubs/ruby/BuiltinPackageDownloader.scala
@@ -43,7 +43,7 @@ case class RubyType(name: String, methods: List[RubyMethod], fields: List[RubyFi
   * @param rubyVersion
   *   \- Ruby version to fetch dependencies for
   */
-class BuiltinPackageDownloader(format: OutputFormat.Value = OutputFormat.zip) {
+class BuiltinPackageDownloader(outputDir: String, format: OutputFormat.Value = OutputFormat.zip) {
   private val logger: Logger = LoggerFactory.getLogger(this.getClass)
 
   private val CLASS    = "class"
@@ -52,7 +52,7 @@ class BuiltinPackageDownloader(format: OutputFormat.Value = OutputFormat.zip) {
   private val browser = JsoupBrowser()
   private val baseUrl = s"https://ruby-doc.org/3.3.0"
 
-  private val baseDir = "src/main/resources/ruby/builtin_types"
+  private val baseDir = s"$outputDir/ruby_builtin_types"
 
   // Below unicode value calculated with: println("\\u" + Integer.toHexString('→' | 0x10000).substring(1))
   // taken from: https://stackoverflow.com/questions/2220366/get-unicode-value-of-a-character
@@ -138,8 +138,10 @@ class BuiltinPackageDownloader(format: OutputFormat.Value = OutputFormat.zip) {
     }
 
     logger.debug("[Ruby]: Zipping builtin-type dir")
-    dir.zipTo(destination = File(s"${baseDir}.zip"))
-    dir.delete()
+
+    if format == OutputFormat.zip then
+      dir.zipTo(destination = File(s"${baseDir}.zip"))
+      dir.delete()
   }
 
   /** Write RubyTypes to JSON files for debugging a readable format
